Bump log4j library version to 2.17
