+++ docs/standalone-library-instrumentation.md
@@ -15,8 +15,8 @@ that can be used if you prefer that over using the Java agent:
 * [Ktor 1.0](../instrumentation/ktor/ktor-1.0/library)
 * [Ktor 2.0](../instrumentation/ktor/ktor-2.0/library)
 * [Lettuce](../instrumentation/lettuce/lettuce-5.1/library)
-* [Log4j appender](../instrumentation/log4j/log4j-appender-2.16/library)
-* [Log4j thread context](../instrumentation/log4j/log4j-context-data/log4j-context-data-2.16/library-autoconfigure)
+* [Log4j appender](../instrumentation/log4j/log4j-appender-2.17/library)
+* [Log4j thread context](../instrumentation/log4j/log4j-context-data/log4j-context-data-2.17/library-autoconfigure)
 * [Logback](../instrumentation/logback/logback-appender-1.0/library)
 * [MongoDB Driver](../instrumentation/mongo/mongo-3.1/library)
 * [OkHttp](../instrumentation/okhttp/okhttp-3.0/library)
