+++ instrumentation/log4j/log4j-appender-2.17/library/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+  id("otel.library-instrumentation")
+}
+
+dependencies {
+  implementation(project(":instrumentation-appender-api-internal"))
+  implementation(project(":instrumentation-appender-sdk-internal"))
+
+  library("org.apache.logging.log4j:log4j-core:2.17.0")
+
+  testImplementation(project(":instrumentation-appender-sdk-internal"))
+  testImplementation("io.opentelemetry:opentelemetry-sdk-logs")
+  testImplementation("io.opentelemetry:opentelemetry-sdk-testing")
+}
+
+tasks.withType<Test>().configureEach {
+  // TODO run tests both with and without experimental log attributes
+  jvmArgs("-Dotel.instrumentation.log4j-appender.experimental.capture-map-message-attributes=true")
+  jvmArgs("-Dotel.instrumentation.log4j-appender.experimental.capture-context-data-attributes=*")
+}
