+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.17/library-autoconfigure/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+  id("otel.library-instrumentation")
+}
+
+base.archivesName.set("${base.archivesName.get()}-autoconfigure")
+
+dependencies {
+  library("org.apache.logging.log4j:log4j-core:2.17.0")
+
+  testImplementation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-2-common:testing"))
+}
