+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.17/library-autoconfigure/src/main/resources/META-INF/services/org.apache.logging.log4j.core.util.ContextDataProvider
