+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.7/javaagent/build.gradle.kts
@@ -6,7 +6,7 @@ muzzle {
   pass {
     group.set("org.apache.logging.log4j")
     module.set("log4j-core")
-    versions.set("[2.7,2.16.0)")
+    versions.set("[2.7,2.17.0)")
     assertInverse.set(true)
   }
 }
@@ -14,9 +14,9 @@ muzzle {
 dependencies {
   library("org.apache.logging.log4j:log4j-core:2.7")
 
-  testInstrumentation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-2.16:javaagent"))
+  testInstrumentation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-2.17:javaagent"))
 
   testImplementation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-2-common:testing"))
 
-  latestDepTestLibrary("org.apache.logging.log4j:log4j-core:2.15.+") // see log4j-context-data-2.16
+  latestDepTestLibrary("org.apache.logging.log4j:log4j-core:2.15.+") // see log4j-context-data-2.17
 }
