+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/log4j/v2_7/Log4j27InstrumentationModule.java
@@ -23,16 +23,14 @@ public class Log4j27InstrumentationModule extends InstrumentationModule {
 
   @Override
   public ElementMatcher.Junction<ClassLoader> classLoaderMatcher() {
-    // want to cover 2.7 through (and including) 2.15.0 (but not including 2.16.0+)
+    // want to cover 2.7 up to (but not including) 2.17.0
     // - ContextDataInjectorFactory introduced in 2.7
-    // - DefaultArbiter introduced in 2.15.0
-    // - LookupMessagePatternConverter introduced in 2.15.0, removed in 2.16.0
+    // - ConfigurationStrSubstitutor introduced in 2.17.0
     return hasClassesNamed("org.apache.logging.log4j.core.impl.ContextDataInjectorFactory")
         .and(
-            not(hasClassesNamed("org.apache.logging.log4j.core.config.arbiters.DefaultArbiter"))
-                .or(
-                    hasClassesNamed(
-                        "org.apache.logging.log4j.core.pattern.MessagePatternConverter$LookupMessagePatternConverter")));
+            not(
+                hasClassesNamed(
+                    "org.apache.logging.log4j.core.lookup.ConfigurationStrSubstitutor")));
   }
 
   @Override
